Camera resize: resize_projection (#1363)
* option to resize the projection when the viewport gets resized

* Fix example code to use resize_projection

* section default camera resize uses resize_projection

* set default resize_projection to True and adjust example code

* adjust resize_projection in sections

* Fix example platformer

* If viewport is provided but projection is not, then projection will match the viewport instead of getting the size of the window

* Fix mypy error

* Fix mypy error v2

* Added Camera `viewport_to_projection_width_ratio` and `viewport_to_projection_height_ratio`
Fixed `Camera.center`

+++ doc/tutorials/raycasting/step_08.py
@@ -195,7 +195,7 @@ class MyGame(arcade.Window):
                         self.player_sprite.center_y - self.height / 2)
         self.camera_sprites.move_to(position, speed)
 
-    def on_resize(self, width: float, height: float):
+    def on_resize(self, width: int, height: int):
         super().on_resize(width, height)
         self.camera_sprites.resize(width, height)
         self.camera_gui.resize(width, height)
